+++ src/XRoadProvider/AssemblyInfo.fs
@@ -5,15 +5,15 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("XRoadProvider")>]
 [<assembly: AssemblyProductAttribute("XRoadProvider")>]
 [<assembly: AssemblyDescriptionAttribute("Type providers for generating types and service interfaces for XRoad producers.")>]
-[<assembly: AssemblyVersionAttribute("1.0.1")>]
-[<assembly: AssemblyFileVersionAttribute("1.0.1")>]
-[<assembly: AssemblyInformationalVersionAttribute("1.0.1")>]
+[<assembly: AssemblyVersionAttribute("1.0.2")>]
+[<assembly: AssemblyFileVersionAttribute("1.0.2")>]
+[<assembly: AssemblyInformationalVersionAttribute("1.0.2")>]
 do ()
 
 module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "XRoadProvider"
     let [<Literal>] AssemblyProduct = "XRoadProvider"
     let [<Literal>] AssemblyDescription = "Type providers for generating types and service interfaces for XRoad producers."
-    let [<Literal>] AssemblyVersion = "1.0.1"
-    let [<Literal>] AssemblyFileVersion = "1.0.1"
-    let [<Literal>] AssemblyInformationalVersion = "1.0.1"
+    let [<Literal>] AssemblyVersion = "1.0.2"
+    let [<Literal>] AssemblyFileVersion = "1.0.2"
+    let [<Literal>] AssemblyInformationalVersion = "1.0.2"