+++ src/XRoadProvider/CodeDom.fs
@@ -1,16 +1,11 @@
 ﻿module internal XRoad.CodeDom
 
-#if !NET40
-open Microsoft.CodeAnalysis.CSharp
-#endif
-
 open Microsoft.CSharp
 open Microsoft.FSharp.Core.CompilerServices
 open System
 open System.CodeDom
 open System.CodeDom.Compiler
 open System.Diagnostics
-open System.Globalization
 open System.IO
 
 open TypeSchema
@@ -237,13 +232,15 @@ module Arr =
 
 module Compiler =
     /// Builds new assembly for provided namespace.
-    let buildAssembly codeNamespace =
+    let buildAssembly codeNamespace (packages: NuGetPackage) =
         let codeCompileUnit = CodeCompileUnit()
         codeCompileUnit.Namespaces.Add(codeNamespace) |> ignore
         codeCompileUnit.ReferencedAssemblies.Add(typeof<ITypeProvider>.Assembly.Location) |> ignore
         codeCompileUnit.ReferencedAssemblies.Add(typeof<BinaryContent>.Assembly.Location) |> ignore
-        codeCompileUnit.ReferencedAssemblies.Add(typeof<NodaTime.LocalDate>.Assembly.Location) |> ignore
-        codeCompileUnit.ReferencedAssemblies.Add(typeof<Optional.Option>.Assembly.Location) |> ignore
+        if packages.HasFlag NuGetPackage.NodaTime then
+            codeCompileUnit.ReferencedAssemblies.Add(typeof<NodaTime.LocalDate>.Assembly.Location) |> ignore
+        if packages.HasFlag NuGetPackage.Optional then
+            codeCompileUnit.ReferencedAssemblies.Add(typeof<Optional.Option>.Assembly.Location) |> ignore
         codeCompileUnit.ReferencedAssemblies.Add("System.dll") |> ignore
         codeCompileUnit.ReferencedAssemblies.Add("System.Net.dll") |> ignore
         codeCompileUnit.ReferencedAssemblies.Add("System.Numerics.dll") |> ignore
@@ -260,107 +257,6 @@ module Compiler =
             printfn "%A" compilerResults.Errors
         compilerResults.CompiledAssembly
 
-/// Extensions for String module and class.
-[<AutoOpen>]
-module String =
-    let isNullOrEmpty = String.IsNullOrEmpty
-
-#if NET40
-    // http://www.ecma-international.org/publications/files/ECMA-ST/Ecma-334.pdf
-
-    let isLetterCharacter (ch: char) =
-        match CharUnicodeInfo.GetUnicodeCategory(ch) with
-        | UnicodeCategory.UppercaseLetter
-        | UnicodeCategory.LowercaseLetter
-        | UnicodeCategory.TitlecaseLetter
-        | UnicodeCategory.ModifierLetter
-        | UnicodeCategory.OtherLetter
-        | UnicodeCategory.LetterNumber -> true
-        | _ -> false
-
-    let isCombiningCharacter (ch: char) =
-        match CharUnicodeInfo.GetUnicodeCategory(ch) with
-        | UnicodeCategory.NonSpacingMark
-        | UnicodeCategory.SpacingCombiningMark -> true
-        | _ -> false
-
-    let inline private isDecimalDigitCharacter (ch: char) = CharUnicodeInfo.GetUnicodeCategory(ch) = UnicodeCategory.DecimalDigitNumber
-    let inline private isConnectingCharacter (ch: char) = CharUnicodeInfo.GetUnicodeCategory(ch) = UnicodeCategory.ConnectorPunctuation
-    let inline private isFormattingCharacter (ch: char) = CharUnicodeInfo.GetUnicodeCategory(ch) = UnicodeCategory.Format
-    let inline private isUnderscoreCharacter (ch: char) = ch = '_'
-    let inline private isIdentifierStartCharacter (ch: char) = isLetterCharacter ch || isUnderscoreCharacter ch
-
-    let private isIdentifierPartCharacter (ch: char) =
-        isLetterCharacter ch || isDecimalDigitCharacter ch || isConnectingCharacter ch || isCombiningCharacter ch || isFormattingCharacter ch
-
-    let private isValidIdentifier (name: string) =
-        if name |> isNullOrEmpty then false else
-        if isIdentifierStartCharacter name.[0] |> not then false else
-        Array.TrueForAll(name.ToCharArray() |> Array.skip 1, Predicate(isIdentifierPartCharacter))
-#else
-    let private isIdentifierPartCharacter = SyntaxFacts.IsIdentifierPartCharacter
-    let private isIdentifierStartCharacter = SyntaxFacts.IsIdentifierStartCharacter
-    let private isValidIdentifier = SyntaxFacts.IsValidIdentifier
-#endif
-
-    /// Joins sequence of elements with given separator to string.
-    let join (sep: string) (arr: seq<'T>) = String.Join(sep, arr)
-
-    type String with
-        /// Converts given XML namespace to class name.
-        member this.ToClassName() =
-            // Remove `http://` prefix from namespace if present.
-            let str =
-                match this.StartsWith("http://") with
-                | true -> this.Substring(7)
-                | _ -> this
-            // Remove special symbols from class name.
-            let className =
-                str.Split('/')
-                |> Array.map (fun p ->
-                    p.Split('.')
-                    |> Array.map (fun x -> CultureInfo.InvariantCulture.TextInfo.ToTitleCase(x.ToLower()).Replace("-", ""))
-                    |> join "")
-                |> join "_"
-            // Check validity of generated class name.
-            if not (isValidIdentifier className) then failwithf "invalid name %s" className
-            className
-        member this.GetValidPropertyName() =
-            let propertyName = Text.StringBuilder()
-            if not (isIdentifierStartCharacter this.[0]) then propertyName.Append("_") |> ignore
-            this.ToCharArray() |> Array.iter (fun c ->
-                if isIdentifierPartCharacter c then propertyName.Append(c) |> ignore
-                elif propertyName.[propertyName.Length - 1] <> '_' then propertyName.Append('_') |> ignore
-                else ())
-            let fixedName = propertyName.ToString()
-            if not (isValidIdentifier fixedName) then failwithf "Invalid property name `%s`." fixedName
-            fixedName
-
-/// Type abstraction for code generator.
-type RuntimeType =
-    /// Simple types that are presented with system runtime types.
-    | PrimitiveType of Type
-    /// Types that are provided by generated assembly.
-    | ProvidedType of CodeTypeDeclaration * string
-    /// Types that represent collection or array of runtime type.
-    | CollectionType of RuntimeType * string * SchemaTypeDefinition option
-    /// Binary content types are handled separately.
-    | ContentType
-    /// Get type name reference for this instance.
-    member this.AsCodeTypeReference(?readonly, ?optional): CodeTypeReference =
-        let readonly = match readonly with Some(true) -> "readonly " | _ -> ""
-        let ctr =
-            match this with
-            | PrimitiveType(typ) -> CodeTypeReference(typ)
-            | ProvidedType(_,name) -> CodeTypeReference(readonly + name)
-            | CollectionType(typ,_,_) -> CodeTypeReference(typ.AsCodeTypeReference(), 1)
-            | ContentType -> CodeTypeReference(typeof<BinaryContent>)
-        match optional with
-        | Some(true) ->
-            let optionalType = CodeTypeReference(typedefof<Optional.Option<_>>)
-            optionalType.TypeArguments.Add(ctr) |> ignore
-            optionalType
-        | _ -> ctr 
 
 /// Create property with backing field.
 let createProperty<'T> name doc (ownerType: CodeTypeDeclaration) =
@@ -376,28 +272,3 @@ let createProperty<'T> name doc (ownerType: CodeTypeDeclaration) =
     ownerType
     |> Cls.addMember backingField
     |> Cls.addMember property
-
-/// Add property to given type with backing field.
-/// For optional members, extra field is added to notify if property was assigned or not.
-let addProperty (name : string, ty: RuntimeType, isOptional) (owner: CodeTypeDeclaration) =
-    let fixedName = name.GetValidPropertyName()
-    let f = Fld.createRef (ty.AsCodeTypeReference(optional=isOptional)) (fixedName + "__backing")
-            |> Fld.describe Attributes.DebuggerBrowsable
-    let p = Prop.createRef (ty.AsCodeTypeReference(optional=isOptional)) fixedName
-            |> Prop.setAttr (MemberAttributes.Public ||| MemberAttributes.Final)
-            |> Prop.addGetStmt (Stmt.ret (Expr.this @=> f.Name))
-            |> Prop.addSetStmt (Stmt.assign (Expr.this @=> f.Name) (Prop.setValue))
-    owner |> Cls.addMember(f) |> Cls.addMember(p) |> ignore
-    p
-
-let addContentProperty (name: string, ty: RuntimeType, useXop) (owner: CodeTypeDeclaration) =
-    let name = name.GetValidPropertyName()
-    Fld.createRef (ty.AsCodeTypeReference(true)) (sprintf "%s { get; private set; } //" name)
-    |> Fld.setAttr (MemberAttributes.Public ||| MemberAttributes.Final)
-    |> Fld.describe (Attributes.xrdElement None None None false true useXop)
-    |> Fld.addTo owner
-    |> ignore
-    Ctor.create()
-    |> Ctor.addParamRef (ty.AsCodeTypeReference()) "value"
-    |> Ctor.addStmt (Stmt.assign (Expr.this @=> name) (!+ "value"))
-    |> Ctor.addTo owner