+++ src/XRoadProvider/Commons.fs
@@ -808,3 +808,92 @@ module internal MultipartMessage =
             let content = new MemoryStream()
             stream.CopyTo(content)
             (upcast content, [])
+
+[<Flags>]
+type internal NuGetPackage =
+    | None     = 0b00
+    | NodaTime = 0b01
+    | Optional = 0b10
+    | All      = 0b11
+
+/// Extensions for String module and class.
+[<AutoOpen>]
+module String =
+#if !NET40
+    open Microsoft.CodeAnalysis.CSharp
+#endif
+
+    open System.Globalization
+
+    let isNullOrEmpty = String.IsNullOrEmpty
+
+#if NET40
+    // http://www.ecma-international.org/publications/files/ECMA-ST/Ecma-334.pdf
+
+    let isLetterCharacter (ch: char) =
+        match CharUnicodeInfo.GetUnicodeCategory(ch) with
+        | UnicodeCategory.UppercaseLetter
+        | UnicodeCategory.LowercaseLetter
+        | UnicodeCategory.TitlecaseLetter
+        | UnicodeCategory.ModifierLetter
+        | UnicodeCategory.OtherLetter
+        | UnicodeCategory.LetterNumber -> true
+        | _ -> false
+
+    let isCombiningCharacter (ch: char) =
+        match CharUnicodeInfo.GetUnicodeCategory(ch) with
+        | UnicodeCategory.NonSpacingMark
+        | UnicodeCategory.SpacingCombiningMark -> true
+        | _ -> false
+
+    let inline private isDecimalDigitCharacter (ch: char) = CharUnicodeInfo.GetUnicodeCategory(ch) = UnicodeCategory.DecimalDigitNumber
+    let inline private isConnectingCharacter (ch: char) = CharUnicodeInfo.GetUnicodeCategory(ch) = UnicodeCategory.ConnectorPunctuation
+    let inline private isFormattingCharacter (ch: char) = CharUnicodeInfo.GetUnicodeCategory(ch) = UnicodeCategory.Format
+    let inline private isUnderscoreCharacter (ch: char) = ch = '_'
+    let inline private isIdentifierStartCharacter (ch: char) = isLetterCharacter ch || isUnderscoreCharacter ch
+
+    let private isIdentifierPartCharacter (ch: char) =
+        isLetterCharacter ch || isDecimalDigitCharacter ch || isConnectingCharacter ch || isCombiningCharacter ch || isFormattingCharacter ch
+
+    let private isValidIdentifier (name: string) =
+        if name |> isNullOrEmpty then false else
+        if isIdentifierStartCharacter name.[0] |> not then false else
+        Array.TrueForAll(name.ToCharArray() |> Array.skip 1, Predicate(isIdentifierPartCharacter))
+#else
+    let private isIdentifierPartCharacter = SyntaxFacts.IsIdentifierPartCharacter
+    let private isIdentifierStartCharacter = SyntaxFacts.IsIdentifierStartCharacter
+    let private isValidIdentifier = SyntaxFacts.IsValidIdentifier
+#endif
+
+    /// Joins sequence of elements with given separator to string.
+    let join (sep: string) (arr: seq<'T>) = String.Join(sep, arr)
+
+    type String with
+        /// Converts given XML namespace to class name.
+        member this.ToClassName() =
+            // Remove `http://` prefix from namespace if present.
+            let str =
+                match this.StartsWith("http://") with
+                | true -> this.Substring(7)
+                | _ -> this
+            // Remove special symbols from class name.
+            let className =
+                str.Split('/')
+                |> Array.map (fun p ->
+                    p.Split('.')
+                    |> Array.map (fun x -> CultureInfo.InvariantCulture.TextInfo.ToTitleCase(x.ToLower()).Replace("-", ""))
+                    |> join "")
+                |> join "_"
+            // Check validity of generated class name.
+            if not (isValidIdentifier className) then failwithf "invalid name %s" className
+            className
+        member this.GetValidPropertyName() =
+            let propertyName = Text.StringBuilder()
+            if not (isIdentifierStartCharacter this.[0]) then propertyName.Append("_") |> ignore
+            this.ToCharArray() |> Array.iter (fun c ->
+                if isIdentifierPartCharacter c then propertyName.Append(c) |> ignore
+                elif propertyName.[propertyName.Length - 1] <> '_' then propertyName.Append('_') |> ignore
+                else ())
+            let fixedName = propertyName.ToString()
+            if not (isValidIdentifier fixedName) then failwithf "Invalid property name `%s`." fixedName
+            fixedName